+++ src/components/flex/flex.test.tsx
@@ -1,0 +1,7 @@
+import { Flex } from '.';
+
+describe('Flex', () => {
+  it('is truthy', () => {
+    expect(Flex).toBeTruthy();
+  });
+});